requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
cryptography
Flask-WTF
WTForms
python-dotenv
gunicorn
Flask-Migrate